+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Formalise que_scheduler_check_job_exists function change [#334](https://github.com/hlascelles/que-scheduler/pull/335)
 - Remove Ruby 2.5, Ruby 2.6 and Que 0.12.x support [#334](https://github.com/hlascelles/que-scheduler/pull/334)
 
 ## 4.2.2 (2022-02-23)